requirements.txt
python-dotenv
redis
pymongo
pydantic
pytest-django
django_db
djangorestframework
django-cors-headers
django-celery-results
django-celery-beat
Django
celery
